requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pymongo
pytesseract
opencv-python
pillow
numpy
scikit-learn
requests
deepface
python-dotenv
